Encode Wizard's state in URL ('Permalink' support)
Save Wizard's state as base64 to URL, so that reloading the page no
longer clears everything already set in the Wizard.
This also allows sharing URL between people with the exact same
environment.
To clear the saved state, simply remove everything from the URL's `d`
parameter and reload the page.
This is especially useful for debugging, could be disabled for
production if not needed.
This feature must be viewed as a BETA version, since there may be
issues if the encoded data is no longer available in Wizard's actual
data.