Fix dropdown list position
Occasionally the dropdown list would appear to the left of its
corresponding button by a few pixels.
To reproduce this issue open a dropdown that doesn't fit in the page
(i.e. days of the month in the schedule). The dropdown list would also
snap to the correct position at the next DOM update.