Increase authorization token cookie expiration
This, actually, doesn't do anything currently. It just means that if the
Keystone authorization token expiration time is increased, the UI client
will also stay authorized in that time.
Currently the Keystone token expires in 1 hour. The UI client can stay
authorized up to 30 days if the Keystone token expiration time is
increased.

This also includes a small code documentation of the UI authorization
process.