Load destination options values in wizard
After the destination options schema is loaded, load the destination
options values which complement the schema. For example, if using the
schema we render a text input for a string field, after loading the
destination options, we may have a dropdown instead with available
values for that field and a default value selected.