Improve duplicated items in the dropdown component
For each duplicated item in a dropdown list, display its ID value bellow
its label, this makes it much more easier to understand what's actually
selected. No changes should be seen if there're no duplicated items.

For a quick test of this feature, create an endpoint with the same name
as an existing one, go to Wizard's source or target selection page and
open the dropdown containing the duplicated endpoint names.