Refactor the process of switching a project
With this, changing the project is handled in on place only, the
`PageHeader` component, instead of being handled in every place that
uses `PageHeader` component.
Also, `Wait.for` function is replaced by an improved switch project
promise, which should gain us a small performance boost while waiting
for the project to change.