Improve 'mobx' performance by adding `@observer`
Adding `@observer` to all React components can have a great positive
impact on performance.
This is due to the 'mobx' architecture which can re-render only the
components which actually use the observable data. By not declaring a
component `@observer`, the parent component may also have to re-render
when an observable data changes.

An official recommendation can be found here:
https://github.com/mobxjs/mobx-react#which-components-should-be-marked-with-observer