Use password for OCI `private_key_passphrase`
Renders the OCI's `private_key_passphrase` field as a password input in
endpoint new / edit modal and as a password value in endpoint details
page.
Also adds the ability to specify custom password field names in the
`EndpointStore` file.
If the password field names must be configurable, the array should be
moved in the config file.