Create e2e tests for users and projects management
The tests are run only if users and projects management is enabled in
the `src/config.js` file.

This commit also fixes an issue where removing a user with multiple
roles from a project (from the project details page) may cause the user
to still be a member of that project (visible only after a page refresh).