Run a cleanup command before running tests
The cleanup will appear as a regular test (so that it can also be
easily run manually and cleanup errors can be seen nicely) and will
delete all the replicas, migrations, endpoints, users created by Cypress
and projects created by Cypress.
The cleanup will happen only if the current user is 'cypress'.

Also includes and updated version of Cypress: 3.0.1 -> 3.0.2.