Cypress `wait` commands refactor, Electron support
Wait in parallel for XHR requests responses, waiting for them in
sequence causes issues if the first request in a sequence takes a long
time to respond.
Added support for an older Chrome version used by Cypress when using its
Electron browser (useful when running Cypress from CLI in a CI
solution).
Includes an updated version of Cypress.