Don't use the label dictionary for all enum values
Enum values should not be transformed by the dictionary in order to be
displayed.
There's a new property in the dictionary which holds all the enum fields
which are an exception to this rule (currently there's only one such
field: 'port_reuse_policy', which means that the enums of this field
will be transformed by the dictionary).