Refactor all `null` / `undefined` checks
In Javascript there's no need to check for both `null` and `undefined`
if the equality operator (`==`) is used instead of the identity operator
(`===`).
The rule (according to default 'eslint' and 'flow' specs) is that the
`===` operator should always be used with one exception, which is when
checking for `null` / `undefined` since `== null` will only coerce a
value to null only if the value is null or undefined.