Cleanup React components constructor
In most situations the constructor does not need to be overridden for
React components. Function binding can be done using an ES6 decorator
and state initialisation can be done outside the constructor.

Also includes an updated React linting package and rules so that the
order / sorting of the React components' functions and variables are
improved.