Improve Azure endpoint fields rendering UX
'Subscription ID' is now a required field.

'Tenant' is a dynamically required field: it is required only for
'CustomCloud' cloud profile and for 'Service Principal Credentials'
login method.

Fixes an issue where 'Tenant' is not visible in 'CustomCloud' cloud
profile.

'CustomCloud' fields are now sorted: the required fields are first, then
the fields are sorted alphabetically.