Remove references to API's migration 'info' field
The structure or existence of the 'info' field in a replica or a
migration can't be trusted. An alternative method is used to retrieve
the VM instances connected to a specific network, which involves making
an API call for each instance of a replica or a migration.

This affects only the replica and migration details page and the
following changes are visible:

 - a 'Loading...' label is display in the 'Connected VMs' field of the
network table while the migration / replica instances details are
loading.

 - an error message may be shown if there are issues retrieving the
instances details.