Remove unused code from previous notification sys.
Persisted notification are no longer a thing, so the code pertaining to
that is no longer needed.

Previously a regular notification could be made 'persisted', using
`persist: true` option, by saving it and displaying it in 'Bell'
component dropdown.

Currently, the 'Bell' component dropdown displays a different set of
predefined notifications.