Use a javascript source map for production builds
This will allow the user to see the original file name and the source
code line number in the case of a Javascript error when using browser's
developer tools.

Note that the full source code can also be enabled in the developer
tools if needed by changing `nosources-source-map` to `source-map` in
`webpack.config` before building.