Improve layout of instances list
When there are multiple instances in a replica / migration, show each
instance in a separate line, instead of listing them comma-separated for
a much better reading experience.

The layout is also kept when copying to clipboard.

The same layout is also applied to the list of migrations / replicas
where an endpoint is used (in the endpoint details page).