Fix dropdown labels layout if label is too long
Split the dropdown list item label on multiple rows if label doesn't fit
in a single row.

Previously, the label was split on multiple rows but only if it
contained multiple words. For example 'A very long label item' would get
multiple rows with no issues, but 'averylonglabelitem' would cause
layout issues.
So, this affects, for example, items with long IDs as labels (see OCI
provider).