Load instances in chunks in the background
Instead of loading a new page of instances on next page click, load all
pages in the background one after one.

Searching is done on the server only if the instances are not completely
loaded in the background.

The UI shows the number of currently loaded pages and indicates whether
all pages have been completely loaded.