Allow Openstack with empty connection info
This can be enabled in `config.js` (`useCurrentCredentialsForOpenstack`).

Enabling this option makes all Openstack fields non-required and skips
the step where the endpoint connection info is added to Barbican secrets.

So, if no connection info is filled when adding a new Openstack
endpoint, an empty object `{}` will be sent as the connection info body.