Remove `#` symbol from production URL routes
Application URLs no longer have `#` (hashtags) when building the
application for production: `yarn build` and `node server`.

All URLs links and redirects now use React Router API (based on HTML5
history API) to make sure that the pages don't refresh when hashtags are
not used. Previously, `window.location.href` was used for redirects
which causes a page refresh when hashtags are not used.