Fix an issue where fields use one column layout
This is reproducible when trying to recreate a migration with OCI as a
source and then zooming out the browser.

This happened because, by default, the size of the browser is taken into
consideration when deciding between 1 or 2 columns.
This works for the wizard options fields, but in this case the max
height of the modal does not depend on the size of the browser.