Allow option to type user domain on login page
The user domain input can be toggled on / off in config file.

The last user domain used is saved in local storage, after a successful
login. This means you can disable the input anytime after the first
login.

A default value is used the first time the user loads the page.