Use custom sorting for provider names
Use a custom sorting algorithm based on a list of sorting priorities to
preferentially sort the provider names.

Used by `ChooseProvider` modal and when choosing the source and target
endpoints in the Wizard page.

You can check `providerSortPriority` in `config.js` to update the list.
Refreshing the browser refreshes it.