Update storage mapping default behaviour
Auto select the server's storage config default in 'Default Storage'
field, found in Wizard target options and Edit Replica target options.

When creating a replica, don't send any storage mapping data to the
server, if its target mapping is 'Default'.

When editing a replica, if a target storage mapping is changed to
'Default', send the 'Default Storage' as the target mapping.

In Wizard Summary page, show only the storage mappings that are actually
sent to the server.