Dynamically detect if a provider has env options
Dynamically detect if a provider can make `destination-options` or
`source-options` API calls to fill the destination / source options
schema. For this, we use Coriolis API's `constants.py` info.

Previously, the list of providers which support these calls was stored
in `config.js`. Note that we still need to store the providers which can
make additional destination / source API calls with a set of field
values, in the config file.