Refactor Wizard and Endpoint options fields
There is no more Wizard Options filed and Endpoint field components. One
component (`FieldInput`) is now used throughout the application, with
the ability to specify the type of layout to be used: 'modal' (formerly
`EndpointField`) and 'page' (formerly `WizardOptionsField`).

This allows easily switching between the two layouts and prepares the
work for switching the layout in Edit Replica modal.