Improve options loading UX
The first time the options are loading, show the loading animation for
the entire page.

For subsequent options loading, show disabled loading for all fields
except for the fields which are required in `env` parameter of the
options calls.

This means that while the options are being loaded, users can still
change those fields, in which case, the previous options call is
canceled and a one is started.