Use a configurable list of password fields
You can now use the `config.js` to specify a list of password fields,
which will be used to rendered inputs as password inputs. Applicable to
options and endpoints fields.

If an option or endpoint field contains the 'password' keyword, the
fields will be rendered as password by default (no need to add it to the
config list).