Display options names instead of IDs
Options names are shown in Wizard Summary page and in Replica /
Migration Details pages, where only the IDs were displayed previously.

For the Wizard Summary page, no extra calls are necessary to display the
option names instead of IDs. For the Replica / Migration Details page,
an extra API call is required, although the call's response is cached
for 15 minutes.

A small loading icon is displayed in the Replica / Migration Details
page, next to the 'Properties' label, while the options names are being
loaded.