Improve and fix admin pages loading UX
Improve the loading UX when the user goes directly to an admin page
(ex.: `/projects`):

- Show 'Checking permissions...' message while waiting for the API to
confirm or deny the user's access to the page.
- Show 'User doesn't have permissions to view this page' message if
that's the case.

Previously, a simple 'Not Found' message has be shown instead.

Fixes some instances where 'Not Found' message was shown when going
directly to an admin page, even if the user has permissions to that page.