Add dictionary support for fields with same name
Previously, if there were multiple fields with the same name, the label
for the first loaded field would be used for all the rest.

With this commit, you can have fields with the same name and different
labels as long as they belong to different providers or they have
different directions (i.e. one is for destination, one is for source).