Fix Safari password save prompt issue in Wizard
If there's a filled-in password field in the Wizard, Safari prompts the
user to store it in keychain every time any input is changed, due to the
Wizard's URL changing on each input change.

As a temporary workaround, the Wizard URL no longer changes for Safari,
meaning the Wizard page can no longer be refreshed while keeping all the
filled values and can no longer be bookmarked.