Preserve skip_os_morphing when recreating migr.
The `skip_os_morphing` field is now shown when recreating a migration
created from scratch.

The default value of the skip os morphing field is deduced by finding a
migration task of type 'OS_MORPHING'.

This default value is used for recreating a migration created from
scratch and for recreating a migration created from a replica.