Fix some dest. / source options values missing
Fixes an issues where, for some providers, the API may not return the
values for fields sent in a previous dest. / source options call.

For example, when requesting destination options for Azure, while
sending the location and resource group as parameters, the list of
locations and resource groups are no longer returned. As opposed to
making the destination options call for Azure the first time, with no
parameters, in which case, the list of locations and resource groups are
returned.

This means that 2 calls must be made each time to retrieve the full list
of options.