Better handling of multi. replicas creation errors
Now, when creating multiple replicas / migrations, there's a graceful
handling of creation errors.

Along with the error messages for each failed replica creation, a
summary alert is also shown with the total number of created and failed
replicas.

If there's at least 1 error, the user is no longer redirected from the
Wizard page.