Check for replica disks when deleting a replica
When deleting a replica, the confirmation alert shows whether the
replica has replica disks.
If the replica has replica disks, an option to delete those disks is
presented.

The same flow is kept when deleting multiple replicas. In this case,
'delete replica disks' is executed just for the replicas which actually
have replica disks and not for all the selected replicas.