Disable update / create while options are loading
This disables the update button when editing a replica and the create
button when recreating a migration while options (source, destination,
network and storage) are being loaded.

Previously, the button would be disabled only if the user was on the
options panel for which the options was currently being loaded.

There's also now a loading indicator for each of the 4 panels, so the
user is informed which panel is still waiting for its options.