Show loading button in Wizard Options
While the first set of options from 'destination-options' API call are
being loaded, a big Loading animation is displayed.

While the second set of options are being loaded, if the user is on
'Simple' tab, he may not notice them loading and thus may not know why
the 'Next' button is disabled.

This commit fixes this issue by display a 'Loading ...' animated button
while the second set of options are being loaded, instead of the
disabled 'Next' button.