Add support for data restructuring in API lists
The API no longer returns everything (like executions and tasks) for
each item when listing replicas or migrations.

The executions are only loaded when opening a replica.
The tasks are loaded only for the selected execution, or in the case of
a migration, when opening a migration.

Some components relied on execution data being present when listing
replicas and migrations. Those components have been rewritten to use
alternative data available in the lists. Among the affected components,
the important ones are the notifications, the Timeline module inside the
Dashboard page, the replicas and migrations list item renderer.