Add Coriolis Minion Pools operations support
The users can view and manage the Coriolis Minion Pools similar to the
other Coriolis resources (Replicas, Migrations, Endpoints etc.). The
Minion Pools can be listed, created, updated, deleted and certain
actions can be applied. Their executions view is similar to that of the
replicas and migrations.

The Minion Pools are also integrated with Replicas and Migrations. When
creating a new one, the supported source and target endpoints have their
associated minion pools listed and a minion pool can be set.

Each instance of a Replica / Migration may be mapped to a supported
target endpoint's Minion Pool.

This Replica / Migration Minion Pool support is available throughout
the UI where their associated operations are available, like editing
replicas and migrations and creating a migration from replica.