Fix Options fields changing in a confusing way
This issue is visible when changing the Openstack source 'Replica
Transfer Mechanism' field, which causes other fields to appear /
disappear, which is confusing, especially since the layout of the fields
also changes from vertical to horizontal.

As a UX hot fix, the layout of the options will always be horizontal and
a small animation will highlight the appearance of new fields.