Add support for duplicated instance names
When displaying instances, a unique value will also be displayed (which
may be either `instance_name` or `id`).

That same unique value will also be used for all API requests where an
instance needs to be provided.

Previously, the unique value sent was `instance_name`, if available, or
just `name`, if not.