Use URI Encoding to encode Base 64 query string
In rare occasions, when using special characters in a string (ex.: ü),
the Base 64 value contains the '/' character, which is not URL query
string friendly and was previously not URI encoded.

This PR makes sure that every time a Base 64 value is used in a query
string, it's also URI encoded to make sure '/' character is not part of
the query string.

Please note that the server must decode (deqoute) the query string
before decoding its Base 64 value.