Special handling for Coriolis Regions export
When exporting an endpoint which has Coriolis Regions mapped, export
their names instead of their IDs.

When importing an endpoint try to match the regions by name.

If there are 1 or more regions that couldn't be found by name, give the
user the options to select a new Coriolis Regions mapping.