Show Setup page when launching for the first time
When the app is launched for the first time a Welcome page is shown with
legal agreements and help links.

The EULA can also be viewed in the app's About modal.

This PR also includes code for trial and paid licence request forms and
subsequent flow steps. That code is temporarily disabled.