Add instance with no ID protection
Previously, the UI entered an infinite loop if there was an instance
with no ID and that instance was the last in the chunk of instances
returned for an endpoint.

With this commit, if there are instances with no ID, they are skipped
and an error is shown in the UI with their name.

This commit also lays the groundwork for future instance validation.