Fix endpoint fields reverting to previous values
Sometimes when opening the edit endpoint modal and immediately starting
to edit a field, the field would revert to its initial value. This
doesn’t happen if the user waits a few seconds before starting to edit a
field.

This was caused by a concurrency issue between requests and fixed by no
longer making the requests if they are already in progress.