Move from “Atomic” to a “Modular” design pattern
This PR will just restructure the code into a different folder pattern,
no functionality was added or modified.

Previously, the code followed the “Atomic Design Pattern”, where React
components were divided into 3 groups based on their “scope”: atoms
(where we put our components with the smallest scope like inputs, icons,
fonts etc.), molecules (an intermediate step with components that group
atoms components) and organisms (groups components from the molecules
and atoms groups).

An issue with this pattern, in our application’s case, is that it
doesn’t scale so well, leaving us with a single folder holding a lot of
components with no clear way to avoid it. This also makes it difficult
to find a component by just browsing the folder structure.

In the current and future state of the application, a “Modular” design
pattern is more appropriate by solving the previous issues. The basic UI
components are stored separately, while the other components are grouped
in modules based on their purpose.