Add custom ESLint code style rule for imports
Implements a custom ESLint plugin (`eslint-plugin-coriolis-web`) that
contains a rule which removes the duplicate name found when importing a
React component.

When using the VSCode auto import feature for React components, the
import path contains both the component parent folder and the component
file name. For this project, the component file name is not required
(because it has a `package.json` file, thus defining it as a module) and
it just pollutes the import path.