Add Coriolis Bare Metal Servers support
There's a new page for displaying the list of Coriolis Bare Metal
Servers available in Coriolis Bare Metal Hub appliance. The servers can
be added, deleted, replicated and migrated.

The Coriolis Bare Metal Hub Fingerprint is displayed on the Servers list
page.

The name of the Coriolis Bare Metal Hub is configurable and moddable
with bareMetalEndpointName.

CA_FINGERPRINT must be set as an environment variable with the path to
the Fingerprint file in order for it to be shown in the UI.