Fix default user domain for Keystone UI login
The `defaultUserDomain` option from the config file was previously
ignored if a successful login was already done with a different user
domain in the same browser.

This is fixed by always using the `defaultUserDomain` option when
logging in, if `showUserDomainInput` is false. If `showUserDomainInput`
is true, then the user domain input is shown when logging in and the
behaviour is kept as it was previously, where the last successful user
domain is pre-filled in the user domain input.