Fix UI crash when selecting `null` value
The UI might previously crash when selecting a null value from a
dropdown list in Edit Replica modal. This was caused by updating the
global state of the replica when editing instead of the local state (the
local state here is the changed replica from the Edit Replica modal).

This was fixed by cloning the global state into the local state before
any changes are made.