Increase the width of the Edit Replica modal
The width of the Edit Replica / Recreate Migration modal has been
increased so that it could fit inputs as large as the ones used in the
Wizard options pages.

This happens only if there's available space in the window, otherwise
the smaller (previous) width will be used.

This PR also provides a smaller app bundle size since some duplicate
background images were removed and instead a generic one was created and
is used everywhere where the background image is required. This removal
was needed in order to properly increase the width of the Edit Replica
modal while keeping the background image (used in the modal's title
bar).