Fix updating field to empty value in edit replica
Previously, when editing a replica by completely removing the value from
a field, the empty field was ignored when building the request body with
the updated data.

This is now fixed by using the current state of the replica to determine
whether the updated field value should be sent even if it is empty.