Show loading message in recreate migration modal
This commit updates the behavior of the recreate migration modal to show
a loading message while the migration is being created. Previously, the
modal was instantly dismissed and a notification was shown that the
migration was created, requiring the user to click a redirect button for
redirection. With this change, the modal will remain open and display a
loading message until the migration has finished being created, at which
point it will automatically redirect to the migration status page. This
improves the user experience by providing more information and reducing
the number of steps required to view the migration status.