Fix and refactor data validation for Edit Replica
Fixed an issue where the validation was not working (shown as invalid)
when a required field was part of a UI group, like "Coriolis Backups
Options".
In this case the UI expected to find the data inside that group, but the
value was actually at the top level of the data object as it was stored
in the source or destination env.
The value gets added to the group when the user updates the field, but
otherwise the validation was failing.

Also refactored the validation code to make it more simple, readable and
maintainable.