Fix Recreate Migration not redirecting properly
When recreating a migration from replica, after the migration is
created, the user is redirected to the migration details page. However,
the migration details page is not updated with the new migration and
polling is not started.
This commit fixes that.

Also, a loading spinner is added to the Recreate Migration modal to
indicate that the migration is being recreated.