Fix execution erorr handling
Prior to this fix, upon clicking on `Execute`, the executions page would
get loaded before waiting for a response. If an error would get returned
from the request, it would get redirected to executions page with a loading
circle, without them actually loading. This patch makes sure to wait for the
response before redirecting to the executions page. If an error occurs, it
gets displayed in a pop-up, but the modal never closes.