Automatically log out from inactive user sessions
This patch adds the `IdleTimer` component to the main App component,
which will be configured to time out if the user is inactive on the
app pages. On timeout, the user gets automatically logged out.

This functionality is disabled by default, and can be configured using
the `inactiveSessionTimeout` configuration option.